Move the `description` module out of `matcher_support`.
Currently `description::Description` is in `matcher_support` because it facilitates writing matchers but is not a required part of the interface. In the future, the return types of `Matcher::describe` and `Matcher::explain_match` will be changed to `description::Description`, making it an essential part of the interface rather than a mere support mechanism. Thus it makes sense to put it in the root module rather than in `matcher_support`.
