+++ googletest/src/matcher_support/mod.rs
@@ -19,7 +19,6 @@
 //! matchers.
 
 pub(crate) mod count_elements;
-pub mod description;
 pub(crate) mod edit_distance;
 pub(crate) mod summarize_diff;
 pub(crate) mod zipped_iterator;
